requirements/base.txt
django==4.0
psycopg2-binary==2.9.3
django-environ==0.9.0
djangorestframework-simplejwt
djangorestframework
ipython

#Yet Another Swagger
drf-yasg
